# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Film Library
    description: BiS Service for querying and managing films and actors info
    version: "1.0"
host: localhost:3239
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /actors:
        get:
            tags:
                - SearchService
            summary: getAllActors SearchService
            description: |4-
                **Required security scopes for jwt**:
                  * `api:read`
            operationId: SearchService#getAllActors
            parameters:
                - name: Authorization
                  in: header
                  description: JWT used for authentication
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SearchServiceActorResultResponseCollection'
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "403":
                    description: Forbidden response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        post:
            tags:
                - ActorService
            summary: addActor ActorService
            description: |4-
                **Required security scopes for jwt**:
                  * `api:write`
            operationId: ActorService#addActor
            parameters:
                - name: Authorization
                  in: header
                  description: JWT used for authentication
                  required: false
                  type: string
                - name: AddActorRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/ActorServiceAddActorRequestBody'
                    required:
                        - ActorInfo
            responses:
                "201":
                    description: Created response.
                    schema:
                        type: integer
                        format: int64
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ActorServiceAddActorAlreadyExistsResponseBody'
                        required:
                            - message
                            - id
                "403":
                    description: Forbidden response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /actors/{ActorID}:
        put:
            tags:
                - ActorService
            summary: updateActorInfo ActorService
            description: |4-
                **Required security scopes for jwt**:
                  * `api:write`
            operationId: ActorService#updateActorInfo
            parameters:
                - name: ActorID
                  in: path
                  required: true
                  type: integer
                - name: Authorization
                  in: header
                  description: JWT used for authentication
                  required: false
                  type: string
                - name: UpdateActorInfoRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/ActorServiceUpdateActorInfoRequestBody'
                    required:
                        - ActorInfo
            responses:
                "201":
                    description: Created response.
                "403":
                    description: Forbidden response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        delete:
            tags:
                - ActorService
            summary: deleteActor ActorService
            description: |4-
                **Required security scopes for jwt**:
                  * `api:write`
            operationId: ActorService#deleteActor
            parameters:
                - name: ActorID
                  in: path
                  description: Actor ID
                  required: true
                  type: integer
                - name: Authorization
                  in: header
                  description: JWT used for authentication
                  required: false
                  type: string
            responses:
                "204":
                    description: No Content response.
                "403":
                    description: Forbidden response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
...
definitions:
    ActorInfoRequestBody:
        title: ActorInfoRequestBody
        type: object
        properties:
            ActorBirthdate:
                type: string
                description: YYYY-MM-DD
                example: "2024-03-18"
                pattern: ^\d{4}\-(0[1-9]|1[012])\-(0[1-9]|[12][0-9]|3[01])$
            ActorName:
                type: string
                description: Name of an Actor
                example: Margo Robbie
                pattern: ^.*\S.*$
                maxLength: 32
            ActorSex:
                type: string
                description: Sex of an Actor
                example: F
                pattern: ^(M|F)$
        description: Describes an Actor to be added
        example:
            ActorBirthdate: "2024-03-18"
            ActorName: Margo Robbie
            ActorSex: F
        required:
            - ActorName
            - ActorSex
            - ActorBirthdate
    ActorResultResponse:
        title: 'Mediatype identifier: application/vnd.actor; view=default'
        type: object
        properties:
            ActorBirthdate:
                type: string
                example: Culpa eos.
            ActorID:
                type: integer
                description: Unique ID of an Actor
                example: 239
                format: int64
            ActorName:
                type: string
                example: Qui et consequatur maiores beatae magnam.
            ActorSex:
                type: string
                example: Non asperiores qui doloremque nostrum.
        description: ActorResultResponse result type (default view)
        example:
            ActorBirthdate: Sit est.
            ActorID: 239
            ActorName: Facere cupiditate consequuntur.
            ActorSex: Occaecati et adipisci.
        required:
            - ActorID
    ActorServiceAddActorAlreadyExistsResponseBody:
        title: ActorServiceAddActorAlreadyExistsResponseBody
        type: object
        properties:
            id:
                type: string
                description: ID of existing data
                example: Unde est omnis.
            message:
                type: string
                description: Error message
                example: Entity already exists
        description: Actor already exists
        example:
            id: Sed omnis laudantium aut et repellendus accusamus.
            message: Entity already exists
        required:
            - message
            - id
    ActorServiceAddActorRequestBody:
        title: ActorServiceAddActorRequestBody
        type: object
        properties:
            ActorInfo:
                $ref: '#/definitions/ActorInfoRequestBody'
        example:
            ActorInfo:
                ActorBirthdate: "2024-03-18"
                ActorName: Margo Robbie
                ActorSex: F
        required:
            - ActorInfo
    ActorServiceUpdateActorInfoRequestBody:
        title: ActorServiceUpdateActorInfoRequestBody
        type: object
        properties:
            ActorInfo:
                $ref: '#/definitions/ActorInfoRequestBody'
        example:
            ActorInfo:
                ActorBirthdate: "2024-03-18"
                ActorName: Margo Robbie
                ActorSex: F
        required:
            - ActorInfo
...
    SearchServiceActorResultResponseCollection:
        title: 'Mediatype identifier: application/vnd.actor; type=collection; view=default'
        type: array
        items:
            $ref: '#/definitions/ActorResultResponse'
        description: GetAllActorsResponseBody is the result type for an array of ActorResultResponse (default view)
        example:
            - ActorBirthdate: Fuga nisi.
              ActorID: 239
              ActorName: Qui ut sunt eveniet non quo laudantium.
              ActorSex: Eos maxime numquam vel dolorem beatae et.
            - ActorBirthdate: Fuga nisi.
              ActorID: 239
              ActorName: Qui ut sunt eveniet non quo laudantium.
              ActorSex: Eos maxime numquam vel dolorem beatae et.
            - ActorBirthdate: Fuga nisi.
              ActorID: 239
              ActorName: Qui ut sunt eveniet non quo laudantium.
              ActorSex: Eos maxime numquam vel dolorem beatae et.
...
securityDefinitions:
    basic_header_Authorization:
        type: basic
        description: |-
            Basic authentication used to authenticate security principal during sign-in

            **Security Scopes**:
              * `api:read`: Read-only access
    jwt_header_Authorization:
        type: apiKey
        description: |-
            Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint.

            **Security Scopes**:
              * `api:read`: Read-only access
              * `api:write`: Read and write access
        name: Authorization
        in: header